web/requirements.txt
# Web Interface Requirements for OMRChecker

# Flask and web server
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# Image processing (already in main project, but needed for web)
pillow>=10.0.0

# Windows TWAIN scanner support
pytwain>=2.1.0; platform_system=="Windows"

# Linux SANE scanner support (optional)
# python-sane>=2.9.1; platform_system=="Linux"

# Werkzeug for secure filename handling
werkzeug>=2.3.0
